chore(rethinkdb): SlackAuth (#10154)
