+++ packages/server/dataloader/primaryKeyLoaderMakers.ts
@@ -10,6 +10,7 @@ import {
   selectMeetingSettings,
   selectOrganizations,
   selectRetroReflections,
+  selectSlackAuths,
   selectSuggestedAction,
   selectTeamPromptResponses,
   selectTeams,
@@ -94,4 +95,8 @@ export const meetingSettings = primaryKeyLoaderMaker((ids: readonly string[]) =>
 
 export const agendaItems = primaryKeyLoaderMaker((ids: readonly string[]) => {
   return selectAgendaItems().where('id', 'in', ids).execute()
+})
+
+export const slackAuths = primaryKeyLoaderMaker((ids: readonly string[]) => {
+  return selectSlackAuths().where('id', 'in', ids).execute()
 })
