+++ packages/server/dataloader/rethinkPrimaryKeyLoaderMakers.ts
@@ -10,7 +10,6 @@ export const meetingMembers = new RethinkPrimaryKeyLoaderMaker('MeetingMember')
 export const newMeetings = new RethinkPrimaryKeyLoaderMaker('NewMeeting')
 export const newFeatures = new RethinkPrimaryKeyLoaderMaker('NewFeature')
 export const notifications = new RethinkPrimaryKeyLoaderMaker('Notification')
-export const slackAuths = new RethinkPrimaryKeyLoaderMaker('SlackAuth')
 export const slackNotifications = new RethinkPrimaryKeyLoaderMaker('SlackNotification')
 export const tasks = new RethinkPrimaryKeyLoaderMaker('Task')
 export const teamInvitations = new RethinkPrimaryKeyLoaderMaker('TeamInvitation')
