+++ packages/server/graphql/private/mutations/removeAllSlackAuths.ts
@@ -1,19 +1,17 @@
 import getRethink from '../../../database/rethinkDriver'
+import getKysely from '../../../postgres/getKysely'
 import {MutationResolvers} from '../resolverTypes'
 
 const removeAllSlackAuths: MutationResolvers['removeAllSlackAuths'] = async () => {
   const r = await getRethink()
-  const now = new Date()
 
   // RESOLUTION
-  const allSlackAuths = await r.table('SlackAuth').filter({isActive: true}).run()
-  const allSlackAuthIds = allSlackAuths.map(({id}) => id)
   const [slackAuthRes, slackNotificationRes] = await Promise.all([
-    r
-      .table('SlackAuth')
-      .getAll(r.args(allSlackAuthIds))
-      .update({botAccessToken: null, isActive: false, updatedAt: now})
-      .run(),
+    getKysely()
+      .updateTable('SlackAuth')
+      .set({botAccessToken: null, isActive: false})
+      .returning('id')
+      .execute(),
     r.table('SlackNotification').delete().run()
   ])
   const data = {
