+++ packages/server/graphql/public/types/SlackNotification.ts
@@ -1,0 +1,10 @@
+import {slackNotificationEventTypeLookup} from '../../../database/types/SlackNotification'
+import {SlackNotificationResolvers} from '../resolverTypes'
+
+const SlackNotification: SlackNotificationResolvers = {
+  eventType: ({event}) => {
+    return slackNotificationEventTypeLookup[event]
+  }
+}
+
+export default SlackNotification
