--- packages/server/graphql/types/SlackIntegration.ts
@@ -1,8 +1,0 @@
-import {GraphQLObjectType} from 'graphql'
-
-const SlackIntegration = new GraphQLObjectType({
-  name: 'SlackIntegration',
-  fields: {}
-})
-
-export default SlackIntegration
