--- packages/server/graphql/types/SlackNotificationEventEnum.ts
@@ -1,16 +1,0 @@
-import {GraphQLEnumType} from 'graphql'
-
-const SlackNotificationEventEnum = new GraphQLEnumType({
-  name: 'SlackNotificationEventEnum',
-  description: 'The event that triggers a slack notification',
-  values: {
-    meetingStart: {},
-    meetingEnd: {},
-    MEETING_STAGE_TIME_LIMIT_END: {}, // user event
-    MEETING_STAGE_TIME_LIMIT_START: {},
-    TOPIC_SHARED: {},
-    STANDUP_RESPONSE_SUBMITTED: {}
-  }
-})
-
-export default SlackNotificationEventEnum
